+++ src/prob/pf.jl
@@ -61,3 +61,103 @@ function build_mc_pf(pm::_PM.AbstractPowerModel)
         constraint_mc_transformer_power(pm, i)
     end
 end
+
+
+"Constructor for Power Flow in current-voltage variable space"
+function build_mc_pf(pm::_PM.AbstractIVRModel)
+    # Variables
+    variable_mc_bus_voltage(pm, bounded = false)
+    variable_mc_branch_current(pm, bounded = false)
+    variable_mc_transformer_current(pm, bounded = false)
+    variable_mc_gen_power_setpoint(pm, bounded = false)
+    variable_mc_load_setpoint(pm, bounded = false)
+
+    # Constraints
+    for (i,bus) in ref(pm, :ref_buses)
+        @assert bus["bus_type"] == 3
+        constraint_mc_theta_ref(pm, i)
+        constraint_mc_voltage_magnitude_only(pm, i)
+    end
+
+    # gens should be constrained before KCL, or Pd/Qd undefined
+    for id in ids(pm, :gen)
+        constraint_mc_gen_setpoint(pm, id)
+    end
+
+    # loads should be constrained before KCL, or Pd/Qd undefined
+    for id in ids(pm, :load)
+        constraint_mc_load_setpoint(pm, id)
+    end
+
+    for (i,bus) in ref(pm, :bus)
+        constraint_mc_load_current_balance(pm, i)
+
+        # PV Bus Constraints
+        if length(ref(pm, :bus_gens, i)) > 0 && !(i in ids(pm,:ref_buses))
+            # this assumes inactive generators are filtered out of bus_gens
+            @assert bus["bus_type"] == 2
+            constraint_mc_voltage_magnitude_only(pm, i)
+            for j in ref(pm, :bus_gens, i)
+                constraint_mc_gen_power_setpoint_real(pm, j)
+            end
+        end
+    end
+
+    for i in ids(pm, :branch)
+        constraint_mc_current_from(pm, i)
+        constraint_mc_current_to(pm, i)
+
+        constraint_mc_bus_voltage_drop(pm, i)
+    end
+
+    for i in ids(pm, :transformer)
+        constraint_mc_transformer_power(pm, i)
+    end
+end
+
+
+"Constructor for Branch Flow Power Flow"
+function build_mc_pf(pm::AbstractUBFModels)
+    # Variables
+    variable_mc_bus_voltage(pm; bounded=false)
+    variable_mc_branch_current(pm)
+    variable_mc_branch_power(pm)
+    variable_mc_gen_power_setpoint(pm; bounded=false)
+
+    # Constraints
+    constraint_mc_model_current(pm)
+
+    for (i,bus) in ref(pm, :ref_buses)
+        constraint_mc_theta_ref(pm, i)
+
+        @assert bus["bus_type"] == 3
+        constraint_mc_voltage_magnitude_only(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_mc_power_balance(pm, i)
+
+        # PV Bus Constraints
+        if length(ref(pm, :bus_gens, i)) > 0 && !(i in ids(pm,:ref_buses))
+            # this assumes inactive generators are filtered out of bus_gens
+            @assert bus["bus_type"] == 2
+
+            constraint_mc_voltage_magnitude_only(pm, i)
+            for j in ref(pm, :bus_gens, i)
+                constraint_mc_gen_power_setpoint_real(pm, j)
+            end
+        end
+    end
+
+    for i in ids(pm, :branch)
+        constraint_mc_power_losses(pm, i)
+        constraint_mc_model_voltage_magnitude_difference(pm, i)
+        constraint_mc_voltage_angle_difference(pm, i)
+
+        constraint_mc_thermal_limit_from(pm, i)
+        constraint_mc_thermal_limit_to(pm, i)
+    end
+
+    # Objective
+    _PM.objective_min_fuel_cost(pm)
+end