+++ test/opf_bf.jl
@@ -6,7 +6,7 @@
 
     @testset "test linearised distflow opf_bf" begin
         @testset "5-bus lpubfdiag opf_bf" begin
-            result = run_mc_opf_bf(case5, LPUBFDiagPowerModel, ipopt_solver)
+            result = run_mc_opf(case5, LPUBFDiagPowerModel, ipopt_solver)
 
             @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 44880; atol = 1e0)
@@ -18,7 +18,7 @@
 
         @testset "3-bus balanced lpubfdiag opf_bf" begin
             pmd = parse_file("../test/data/opendss/case3_balanced.dss")
-            sol = run_mc_opf_bf(pmd, LPUBFDiagPowerModel, ipopt_solver; make_si=false)
+            sol = run_mc_opf(pmd, LPUBFDiagPowerModel, ipopt_solver; make_si=false)
 
             @test sol["termination_status"] == LOCALLY_SOLVED
             @test isapprox(sum(sol["solution"]["voltage_source"]["source"]["pg"] * sol["solution"]["settings"]["sbase"]), 0.0183456; atol=2e-3)
@@ -27,7 +27,7 @@
 
         @testset "3-bus unbalanced lpubfdiag opf_bf" begin
             pmd = parse_file("../test/data/opendss/case3_unbalanced.dss")
-            sol = run_mc_opf_bf(pmd, LPUBFDiagPowerModel, ipopt_solver; make_si=false)
+            sol = run_mc_opf(pmd, LPUBFDiagPowerModel, ipopt_solver; make_si=false)
 
             @test sol["termination_status"] == LOCALLY_SOLVED
             @test isapprox(sum(sol["solution"]["voltage_source"]["source"]["pg"] * sol["solution"]["settings"]["sbase"]), 0.0214812; atol=2e-3)
@@ -37,7 +37,7 @@
 
     @testset "test linearised distflow opf_bf in diagonal matrix form" begin
         @testset "5-bus lpdiagubf opf_bf" begin
-            result = run_mc_opf_bf(case5, LPUBFDiagPowerModel, ipopt_solver)
+            result = run_mc_opf(case5, LPUBFDiagPowerModel, ipopt_solver)
 
             @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 44880; atol = 1e0)
@@ -46,7 +46,7 @@
 
     @testset "test linearised distflow opf_bf in full matrix form" begin
         @testset "5-bus lpfullubf opf_bf" begin
-            result = run_mc_opf_bf(case5, LPUBFDiagPowerModel, ipopt_solver)
+            result = run_mc_opf(case5, LPUBFDiagPowerModel, ipopt_solver)
 
             @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 44880; atol = 1e0)