+++ test/opf_iv.jl
@@ -4,7 +4,7 @@
     @testset "test IVR opf_iv" begin
         @testset "2-bus diagonal acp opf" begin
             pmd = parse_file("../test/data/opendss/case2_diag.dss")
-            sol = run_mc_opf_iv(pmd, IVRPowerModel, ipopt_solver; make_si=false)
+            sol = run_mc_opf(pmd, IVRPowerModel, ipopt_solver; make_si=false)
 
             @test sol["termination_status"] == LOCALLY_SOLVED
             @test isapprox(sol["objective"], 0.018208969542066918; atol = 1e-4)
@@ -15,7 +15,7 @@
 
         @testset "3-bus balanced acp opf" begin
             pmd = parse_file("../test/data/opendss/case3_balanced.dss")
-            sol = run_mc_opf_iv(pmd, IVRPowerModel, ipopt_solver; make_si=false)
+            sol = run_mc_opf(pmd, IVRPowerModel, ipopt_solver; make_si=false)
 
             @test sol["termination_status"] == LOCALLY_SOLVED
             @test isapprox(sol["objective"], 0.018345004773175046; atol = 1e-4)
@@ -26,7 +26,7 @@
 
         @testset "3-bus unbalanced acp opf" begin
             pmd = parse_file("../test/data/opendss/case3_unbalanced.dss")
-            sol = run_mc_opf_iv(pmd, IVRPowerModel, ipopt_solver; make_si=false)
+            sol = run_mc_opf(pmd, IVRPowerModel, ipopt_solver; make_si=false)
 
             @test sol["termination_status"] == LOCALLY_SOLVED
             @test isapprox(sol["objective"], 0.021481176584287; atol = 1e-4)